JBPM-5914 checks for bad container ids, task instance ids and query param format exceptions

+++ kie-server-parent/kie-server-services/kie-server-services-jbpm/src/main/java/org/kie/server/services/jbpm/admin/UserTaskAdminServiceBase.java
@@ -22,6 +22,7 @@ import java.util.Map;
 import java.util.function.Function;
 import java.util.stream.Collectors;
 
+import org.jbpm.services.api.DeploymentNotFoundException;
 import org.jbpm.services.api.admin.TaskNotification;
 import org.jbpm.services.api.admin.TaskReassignment;
 import org.jbpm.services.api.admin.UserTaskAdminService;
@@ -68,6 +69,8 @@ public class UserTaskAdminServiceBase {
 
     public void addPotentialOwners(String containerId, Number taskId, boolean removeExisting, String payload, String marshallingType) {
         logger.debug("About to unmarshall payload '{}' to map of task inputs", payload);
+        if (context.getContainer(containerId) == null) throw new DeploymentNotFoundException(containerId);
+
         OrganizationalEntity[] entities = convert(containerId, taskId, payload, marshallingType);
 
         userTaskAdminService.addPotentialOwners(taskId.longValue(), removeExisting, entities);
@@ -76,6 +79,7 @@ public class UserTaskAdminServiceBase {
 
     public void addExcludedOwners(String containerId, Number taskId, boolean removeExisting, String payload, String marshallingType) {
         logger.debug("About to unmarshall payload '{}' to map of task inputs", payload);
+        if (context.getContainer(containerId) == null) throw new DeploymentNotFoundException(containerId);
         OrganizationalEntity[] entities = convert(containerId, taskId, payload, marshallingType);
 
         userTaskAdminService.addExcludedOwners(taskId.longValue(), removeExisting, entities);
@@ -84,6 +88,8 @@ public class UserTaskAdminServiceBase {
 
     public void addBusinessAdmins(String containerId, Number taskId, boolean removeExisting, String payload, String marshallingType) {
         logger.debug("About to unmarshall payload '{}' to map of task inputs", payload);
+
+        if (context.getContainer(containerId) == null) throw new DeploymentNotFoundException(containerId);
         OrganizationalEntity[] entities = convert(containerId, taskId, payload, marshallingType);
 
         userTaskAdminService.addBusinessAdmins(taskId.longValue(), removeExisting, entities);
@@ -92,14 +98,18 @@ public class UserTaskAdminServiceBase {
 
     public void removePotentialOwners(String containerId, Number taskId, List<String> orgEntities, boolean isUser) {
         logger.debug("About to remove {} from task {} as potential owners", orgEntities, taskId);
+ 
+        if (context.getContainer(containerId) == null) throw new DeploymentNotFoundException(containerId);
         OrganizationalEntity[] entities = convert(orgEntities, isUser);
-
+    
         userTaskAdminService.removePotentialOwners(taskId.longValue(), entities);
         logger.debug("Potential owners {} removed task {}", entities, taskId);
     }
 
     public void removeExcludedOwners(String containerId, Number taskId, List<String> orgEntities, boolean isUser) {
         logger.debug("About to remove {} from task {} as excluded owners", orgEntities, taskId);
+
+        if (context.getContainer(containerId) == null) throw new DeploymentNotFoundException(containerId);
         OrganizationalEntity[] entities = convert(orgEntities, isUser);
 
         userTaskAdminService.removeExcludedOwners(taskId.longValue(), entities);
@@ -108,6 +118,8 @@ public class UserTaskAdminServiceBase {
 
     public void removeBusinessAdmins(String containerId, Number taskId, List<String> orgEntities, boolean isUser) {
         logger.debug("About to remove {} from task {} as business admins", orgEntities, taskId);
+
+        if (context.getContainer(containerId) == null) throw new DeploymentNotFoundException(containerId);
         OrganizationalEntity[] entities = convert(orgEntities, isUser);
 
         userTaskAdminService.removeBusinessAdmins(taskId.longValue(), entities);
@@ -116,6 +128,8 @@ public class UserTaskAdminServiceBase {
 
     public void addTaskInputs(String containerId, Number taskId, String payload, String marshallingType) {
         logger.debug("About to unmarshall payload '{}' to map of task inputs", payload);
+
+        if (context.getContainer(containerId) == null) throw new DeploymentNotFoundException(containerId);
         Map<String, Object> data = marshallerHelper.unmarshal(containerId, payload, marshallingType, Map.class, new ByTaskIdContainerLocator(taskId.longValue()));
 
         logger.debug("Task input data to be added to a task {} is {}", taskId, data);
@@ -126,6 +140,7 @@ public class UserTaskAdminServiceBase {
     public void removeTaskInputs(String containerId, Number taskId, List<String> inputNames) {
         logger.debug("About to remove task inputs {} from task {}", inputNames, taskId);
 
+        if (context.getContainer(containerId) == null) throw new DeploymentNotFoundException(containerId);
         userTaskAdminService.removeTaskInputs(taskId.longValue(), inputNames.toArray(new String[inputNames.size()]));
         logger.debug("Task inputs {} removed successfully from task {}", inputNames, taskId);
     }
@@ -133,6 +148,7 @@ public class UserTaskAdminServiceBase {
     public void removeTaskOutputs(String containerId, Number taskId, List<String> outputNames) {
         logger.debug("About to remove task outputs {} from task {}", outputNames, taskId);
 
+        if (context.getContainer(containerId) == null) throw new DeploymentNotFoundException(containerId);
         userTaskAdminService.removeTaskOutputs(taskId.longValue(), outputNames.toArray(new String[outputNames.size()]));
         logger.debug("Task outputs {} removed successfully from task {}", outputNames, taskId);
     }
@@ -140,6 +156,7 @@ public class UserTaskAdminServiceBase {
     public String reassignWhenNotStarted(String containerId, Number taskId, String timeExpression, String payload, String marshallingType) {
         logger.debug("About to unmarshall payload '{}' to list of org entities (users/groups)", payload);
 
+        if (context.getContainer(containerId) == null) throw new DeploymentNotFoundException(containerId);
         OrganizationalEntity[] entities = convert(containerId, taskId, payload, marshallingType);
 
         Long id = userTaskAdminService.reassignWhenNotStarted(taskId.longValue(), timeExpression, entities);
@@ -151,6 +168,8 @@ public class UserTaskAdminServiceBase {
 
     public String reassignWhenNotCompleted(String containerId, Number taskId, String timeExpression, String payload, String marshallingType) {
         logger.debug("About to unmarshall payload '{}' to list of org entities (users/groups)", payload);
+
+        if (context.getContainer(containerId) == null) throw new DeploymentNotFoundException(containerId);
         OrganizationalEntity[] entities = convert(containerId, taskId, payload, marshallingType);
 
         Long id = userTaskAdminService.reassignWhenNotCompleted(taskId.longValue(), timeExpression, entities);
@@ -163,6 +182,8 @@ public class UserTaskAdminServiceBase {
 
     public String notifyWhenNotStarted(String containerId, Number taskId, String timeExpression, String payload, String marshallingType) {
         logger.debug("About to unmarshall payload '{}' to EmailNotification (when not started) of task inputs", payload);
+
+        if (context.getContainer(containerId) == null) throw new DeploymentNotFoundException(containerId);
         EmailNotification emailNotification = marshallerHelper.unmarshal(containerId, payload, marshallingType, EmailNotification.class, new ByTaskIdContainerLocator(taskId.longValue()));
         logger.debug("Email notification to be added to a task {} is {}", taskId, emailNotification);
 
@@ -178,6 +199,8 @@ public class UserTaskAdminServiceBase {
 
     public String notifyWhenNotCompleted(String containerId, Number taskId, String timeExpression, String payload, String marshallingType) {
         logger.debug("About to unmarshall payload '{}' to EmailNotification (when not completed) of task inputs", payload);
+
+        if (context.getContainer(containerId) == null) throw new DeploymentNotFoundException(containerId);
         EmailNotification emailNotification = marshallerHelper.unmarshal(containerId, payload, marshallingType, EmailNotification.class, new ByTaskIdContainerLocator(taskId.longValue()));
         logger.debug("Email notification to be added to a task {} is {}", taskId, emailNotification);
 
@@ -194,6 +217,7 @@ public class UserTaskAdminServiceBase {
     public void cancelNotification(String containerId, Number taskId, Number notificationId) {
         logger.debug("About to cancel notification {} from task {}", notificationId, taskId);
 
+        if (context.getContainer(containerId) == null) throw new DeploymentNotFoundException(containerId);
         userTaskAdminService.cancelNotification(taskId.longValue(), notificationId.longValue());
         logger.debug("Notification {} canceled successfully for task {}", notificationId, taskId);
     }
@@ -201,6 +225,7 @@ public class UserTaskAdminServiceBase {
     public void cancelReassignment(String containerId, Number taskId, Number reassignmentId) {
         logger.debug("About to cancel reassignment {} from task {}", reassignmentId, taskId);
 
+        if (context.getContainer(containerId) == null) throw new DeploymentNotFoundException(containerId);
         userTaskAdminService.cancelReassignment(taskId.longValue(), reassignmentId.longValue());
         logger.debug("Reassignment {} canceled successfully for task {}", reassignmentId, taskId);
     }
